feat():员工注册账号，企业
-实现员工、账号、企业的注册，且能互相绑定
-新增registerService

+++ level1/src/main/java/com/yealink/level1/bean/Enterprise.java
@@ -23,9 +23,9 @@ public class Enterprise {
 
     private String name;
 
-    private BigInteger createTime;
+    private long createTime;
 
-    private BigInteger modifyTime;
+    private long modifyTime;
 
 
 }
